Validate that ChapterZip zip code match the state of the chapter
fixes #16

Limiting the choices of the zip code field on ChapterZips proved to be a bit more complicated than we have time to address. Instead, validate after form
submission that the chosen zip code belongs to the
state set on the chapter.

It also adjusts the `__str__` method of Zips to include the state which should make the error easy to spot.